hng9-task-1/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
whitenoise==6.2.0
